==> Spotify Hit Prediction/1_Data Wrangling/1_Wrangling for Billboard Data.ipynb <==
In this notebook I will clean the Billboard data that I collected: no duplicates, no missing values, consistent data types, 

The data for each year is kept in CSV files in the Datasets Folder from the root.

There should be 20 csvs with 100 records each. 

Methodology:

Concatenate all csvs into one large df. I can then check for missing values and duplicates. 


#Import packages
import pandas as pd
import glob

path = 'BillboardData'
all_files = glob.glob(path + "*/.csv")

l1 = []

bill_datasets = ['01','02','03','04','05','06','07','08','09','10','11','12','13','14','15','16','17','18','19','20']

#concatenate all csvs
for year in bill_datasets:
    
    df = pd.read_csv(f'Billboard Data/{year}.csv', index_col=None, header = 0)
    l1.append(df)

billboard_df = pd.concat(l1, axis = 0, ignore_index = True)

billboard_df.head(102)

billboard_df.describe()

#Check for duplicates
duplicates = billboard_df[billboard_df.duplicated(subset=['Artist','Title'], keep=False)]

duplicates.sort_values(by=['Artist', 'Title'])



billboard_df.drop_duplicates(subset = ['Artist','Title'], inplace=True)

billboard_df.info()

We found 124 duplicates and removed them.

dups = billboard_df[billboard_df.duplicated(subset=['Artist','Title'], keep=False)]
dups

billboard_df.describe()



Time to deal with missing values. There was 1 initial missing Artist name (we would have expected 2000 values in the Artist column and found 1999). Serendipitously, upon inspection of the first dataset I found it contained 99 values for Artist and Title. There are therefor only 99 Artist : Title pairs in the 2001 dataset.

However, this still leaves the one Artist that is missing a Title. Let's see what that is..

billboard_df.isna().sum()

#Explore missing values and determine what the issues are

missing = billboard_df[billboard_df['Title'].isna()]

missing

The missing value is actually pretty obvious. It is the song '23' by Mike Will Made-it feat. Miley Cyrus, Wiz Khalifa, etc. 

Lucky the Artist column was so specific! Lets add the value back in to complete the df. 

billboard_df.loc[1239,'Title'] = '23'

billboard_df.describe()

#Found an incorrect Artist name. Need to correct All for you by Janet to all for you by Janet Jackson

billboard_df.loc[2,'Artist'] = 'Janet Jackson'

#replace all instance of Pink, the artist) with the correct spelling: "p!ink"

billboard_df.replace(to_replace = 'Pink', value = 'P!nk',inplace = True)

billboard_df.head()

#Save billboard_df to csv. 

billboard_df.to_csv('BillboardFullDataset.csv')
