==> Spotify Hit Prediction/1_Data Wrangling/3_Flop Songs.ipynb <==
In this notebook I will wrangle the data for the flop songs (non-hits). I will do this using a random name generator that I will feed to a search function. I will order the results by album release date and then pick a song whose URI is NOT in the billboard dataset.

There are two conditions:
1. The song cannot be part of the billboard hit data
2. The song needs to have been released in the same time period (2001-2020)

Let's get started.

I have downloaded a a list of random english words from MIT that is stored in a csv called Random_Words.csv. 

I will randomly pick words from this list that will be the basis for the tracknames of the flop songs

# Import packages

import pandas as pd
import random
import datetime
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
from random import sample
import math

#import csvs with billboard data and random word list
random_words = pd.read_csv('FlopSongsData/Random_Words.csv')
billboard = pd.read_csv('BillboardData/BillboardWithURI.csv')

random_words

# get api key from your .env file

import os

from dotenv import load_dotenv

from pathlib import Path

 

dotenv_path = Path('./SpotifyAPIKey.env')

load_dotenv(dotenv_path=dotenv_path)

 

#load_dotenv()

API_CLIENT_ID = os.getenv('SPOTIFY_CLIENT_ID')
API_CLIENT_SECRET = os.getenv('SPOTIFY_CLIENT_SECRET')
 

Methodology: 


create function to get the random uri from the random word. then get the title and artist from that random uri. store it in a list. the function should look back over the list to make sure there are no matching uris.



URIs = []

#create a random sample of words from the list of words with repetition

for i in range(2001):
    word = random_words.sample(i, replace= True, random_state=42)

type(word)

#Get Flop song URIs and append to the empty list above

def get_uri(rand_word):
    client_credentials_manager = SpotifyClientCredentials(client_id=API_CLIENT_ID,client_secret=API_CLIENT_SECRET)
    sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)
    track = sp.search(q="track:" +f"{rand_word} + year: 2001-2020",type='track',limit=10)
    result = track['tracks']['items']
    
    for items in result:
        uri = items['id']
        dict1 = items['album'] 
        if uri not in billboard['URI']:
            URIs.append(uri)

for lab,row in word.iterrows():
    get_uri(row['words'])

URIs

len(URIs)

#number of unique values in URI flop list and remove duplicates
URI_set = set(URIs)
len(URI_set)

#List of unique Flop song URIs as a set.

URI_list = list(URI_set)

# understanding how the sp.track function works
URI = '74atKkOasLOVzvqB6mYgga'
client_credentials_manager = SpotifyClientCredentials(client_id=API_CLIENT_ID,client_secret=API_CLIENT_SECRET)
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)
track_name = sp.track(URI)

print(track_name)

#define function to get the track names from spotify api
def get_track_name (uri):
    client_credentials_manager = SpotifyClientCredentials(client_id=API_CLIENT_ID,client_secret=API_CLIENT_SECRET)
    sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)
    tracks = sp.track(uri)
    
    track_name = tracks['name']
    return track_name

#test
get_track_name('74atKkOasLOVzvqB6mYgga')

#define function to get the artist names from the spotify api
def get_artist_name(uri):
    client_credentials_manager = SpotifyClientCredentials(client_id=API_CLIENT_ID,client_secret=API_CLIENT_SECRET)
    sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)
    tracks = sp.track(uri)
    
    artist_name = tracks['artists']
    return artist_name[0]['name']

#test
get_artist_name('74atKkOasLOVzvqB6mYgga')

df= pd.DataFrame(URI_list)

df.head()

df = df.rename(columns={0 : 'URI'})

df.head()

#get the titles for all the URIs I collected
for lab,row in df.iterrows():
    df.loc[lab, 'Title'] = get_track_name(row['URI'])

df.head()

#Get the artist names for all the URIs I collected

for lab,row in df.iterrows():
    df.loc[lab, 'Artist'] = get_artist_name(row['URI'])

df.head()

#re-order the 
df= df[['Title', 'Artist', 'URI']]

df.head()

df.to_csv('flop_songs.csv')
