==> Spotify Hit Prediction/1_Data Wrangling/4_Spotipy - Add Features.ipynb <==
In this Notebook I will get the Audio Analysis features for the billboard songs from the Spotify API. The API offers the following musical analysis features for download: 

1. Accosticness
2. Danceability
3. Duration (ms)
4. Energy
5. Instrumentalness
6. Key
7. Liveness
8. Loudness
9. Mode
10. Speechiness
11. Tempo
13. Time Signature
14. Valene

I will be able to use these features and associated target variable (1 or 0 for hit flop) to predict what songs are likely to make it to the top 100 Billboard chart evry year. 


# get api key from your .env file

import os

from dotenv import load_dotenv

from pathlib import Path

 

dotenv_path = Path('./SpotifyAPIKey.env')

load_dotenv(dotenv_path=dotenv_path)


#load_dotenv()

API_CLIENT_ID = os.getenv('SPOTIFY_CLIENT_ID')
API_CLIENT_SECRET = os.getenv('SPOTIFY_CLIENT_SECRET')

 

import pandas as pd
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
import numpy as np


#https://spotipy.readthedocs.io/en/2.19.0/

client_credentials_manager = SpotifyClientCredentials(client_id=API_CLIENT_ID,client_secret=API_CLIENT_SECRET)
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)

#sp.audio_features({'URI'}) gives us a dictionary within a list of the features and their values
#testing how this works

def GetAudioFeatures(URI):
    features = sp.audio_features(URI)
    #df_features = pd.DataFrame.from_dict(features)
    #dataframe = dataframe.merge(df_features, how='inner', left_on='URI', right_on='id')
    return features[0]
 


dict1_ = sp.audio_features('5c1QJAiC3773EwiZLwHf5x')

dict1_

df_t1 = pd.DataFrame.from_dict(dict1_)

df_t1.head()

d = {'URI': ['5c1QJAiC3773EwiZLwHf5x', '5eCkuGPZNl4mISFNPgr3Dd'], 'Artist': ['name1', 'name2']}

df2_test = pd.DataFrame(data=d)

df2_test.head()

#df2_test = GetAudioFeatures(df2_test['URI'], df2_test)

df2_test

#import dataframes

billboard_songs = pd.read_csv('BillboardData/BillboardWithURI.csv', usecols=[1,2,3])

flop_songs = pd.read_csv('FlopSongsData/flop_songs.csv', usecols=[1,2,3])

billboard_songs.info()

flop_songs.info()

billboard_songs.head()

flop_songs.head()

#Apply the function to get the Audio Features to later append them to the DF.

df = billboard_songs['URI'].map(lambda x : GetAudioFeatures(x))

df.head()

#create a dataframe out of the features

df = df.apply(pd.Series)

df.head()

#Adding the features to the billboard song list by merging the two dataframes on the URI

billboard_songs = billboard_songs.merge(df, how='inner', left_on='URI', right_on='id')

#Adding the features to the flog song list

df1 = flop_songs['URI'].map(lambda x : GetAudioFeatures(x))


df1 = df1.apply(pd.Series)

df1.head()

#Adding the features to the flog song list by merging the two dataframes on the URI

flop_songs = flop_songs.merge(df1, how='inner', left_on='URI', right_on='id')

flop_songs.info()

#Set the target variable to 1 for hit songs and 0 for non-hit songs
billboard_songs['Target'] = 1

flop_songs['Target'] = 0

#Concatenating the two dataframes and droping the dupllicates by URI to remove the songs that appear both in the billboard list 
#and the flop list. I will keep the hit songs and remove them from the flop list.

df_con = pd.concat([billboard_songs, flop_songs]).drop_duplicates(subset='URI', keep='first', ignore_index=True)

df_con.head()

df_con.info()

hits=df_con[df_con['Target']==1]

hits.info()

#Remove the unnecessary columns
df_con = df_con.drop(columns = ['type', 'id','uri','track_href', 'analysis_url'])

df_con.info()

n = len(pd.unique(df_con['URI']))
  
print("No. of unique URIs :", 
      n)

#Save data to csv.

df_con.to_csv('FullDataset.csv')
